+++ Project.toml
@@ -17,4 +17,3 @@ Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 attitudeFunctions = "7493c2ef-4a11-4f18-aeab-b8ca9ae8f56a"
-visibilityGroups = "dde82544-3cd7-4f96-a535-ddb3fa36ba29"
